punctuation rule override added. keyboard phrase transformers added
I initially just wanted to override the phrases that were for a particular key presses and keyboard keys using the words.txt but i've begun to realize that some of these key presses are not part of continuous command recognition meaning that you have to say the word press beforehand.

this is why i ended up having to override the punctuation rule for the phrase for the key press back space. I'd like to just say the word backs. However now i'm just realizing as i write this that it might have been better to just do it in the navigation rule or a different rule. I was initially messing with the punctuation role because it was interfering with the word backslash. Actually i would need to find where the back spacing of character is handled using the word clear and i guess it might be in a different rule may be navigation. Hopefully i'll be able to fix this in the next commit.