added the f2 key ccr with a custom rule
the phrase for the f2 key is flash. Realizing that i need to just create one file for all the global ccr keys otherwise i'll just be too many single files and wholly more messy. that's why this file with this new rule it will be the one to move the other commands into