update words.txt to fix misrecognition
so i often have to say the word put inside a sentence when i am dictating and this will cause the dictation to stop and the action of pasting what's in my clipboard to be executed.

instead of having the transform for the word spark to puts, from this point on i will learn to use the  phrase spark for pasting text.

I'm beginning to realize that a lot of these default words in caster that have been selected for the mappings to actions have been selected for good reason