Update ms_word.py, update global, update taskbar
Word: Added commands for read aloud mode.
Global: Added a hacky command for moving a widow to the center of my screen assuming it is snapped to the right side of my screen.
Taskbar: Simplified some of the Extra names so the code paste will better fit in my thesis paper.