Update powershell, notepadplusplus, global rules
Powershell: Removed the "clipboard out" command because it was moved to the CCR rule.

PowershellCCR: Removed "dirrup" because it is in the nonccr rule and moved the remaining exes/commands into the environment variables EXECUTABLES dictionary.

Notepadplusplus: Added a command for toggling the folder as a workspace.

GlobalCCR: Added a command that prints the text ".txt".