Removed window switching section from README
This update removes the detailed instructions and screenshot related to window switching customizations for the Windows 10 taskbar. The README now only contains a brief description of Caster as an extension to the Dragonfly framework for voice command control. No new commands were added in this change.

The reason this was removed is because now that I am in windows 11 I am using different window switching methods which I plan to get into a better state before I explain them.