Add "open with choose" command to Quick Picture Viewer rule
Added a new voice command "open with choose" (Ctrl+Shift+E) to the Quick Picture Viewer rule, allowing users to open the current image with a program of their choice. The command was implemented as a keyboard shortcut that triggers the "Open With" dialog.

Commented out two non-functional commands that were left for reference:
- "open with default" (Ctrl+E)
- "open with paint" (Ctrl+Shift+O)