Fix pip commands for python in command line
This commit modifies the pip and pip install voice commands in the Python rule to invoke pip as a module.

The pip command was changed from pip to -m pip.
The pip install command was changed from pip install to -m pip install.

This makes the commands actually work whereas previously there were not working.